Loader commits in chunks

+++ src/cspro2sql/writer/InsertWriter.java
@@ -3,6 +3,7 @@ package cspro2sql.writer;
 
 import cspro2sql.bean.Dictionary;
 import cspro2sql.bean.Item;
+import cspro2sql.bean.Questionnaire;
 import cspro2sql.bean.Record;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -17,10 +18,14 @@ import java.util.Map;
  */
 public class InsertWriter {
 
-    public static void create(String schema, Dictionary dictionary, Map<Record, List<List<String>>> descr, Statement stmt, StringBuilder script) throws SQLException {
+    public static void create(String schema, Dictionary dictionary, Questionnaire quest, Statement stmt) throws SQLException {
+        create(schema, dictionary, quest, stmt, null);
+    }
+
+    public static void create(String schema, Dictionary dictionary, Questionnaire quest, Statement stmt, StringBuilder script) throws SQLException {
         int id = 0;
         boolean exists = false;
-        for (Map.Entry<Record, List<List<String>>> e : descr.entrySet()) {
+        for (Map.Entry<Record, List<List<String>>> e : quest.getMicrodataSet()) {
             Record record = e.getKey();
             String selectSql = "select ID from " + schema + "." + record.getTableName() + " where ";
             String sql = "insert into " + schema + "." + record.getTableName() + " (";
@@ -82,14 +87,14 @@ public class InsertWriter {
             */
             
             if (exists && !record.isMainRecord()) {
-                script.append("delete from ").append(schema).append(".").append(record.getTableName()).append(" where ").append(record.getMainRecord().getName()).append("=").append(id).append(";\n");
+                if (script!=null) script.append("delete from ").append(schema).append(".").append(record.getTableName()).append(" where ").append(record.getMainRecord().getName()).append("=").append(id).append(";\n");
                 stmt.executeUpdate("delete from " + schema + "." + record.getTableName() +
                         " where " + record.getMainRecord().getName() + "=" + id);
             }
-            script.append(sql).append(";\n");
+            if (script!=null) script.append(sql).append(";\n");
             stmt.executeUpdate(sql);
             if (record.isMainRecord()) {
-                script.append("select last_insert_id()").append(";\n");
+                if (script!=null) script.append("select last_insert_id()").append(";\n");
                 ResultSet lastInsertId = stmt.executeQuery("select last_insert_id()");
                 lastInsertId.next();
                 id = lastInsertId.getInt(1);
